Using App.config instead of other methods of getting configuration.
There are a few I didn’t put in this commit because they’re on lines with other, unrelated changes.  Will merge them in next.

This commit also includes cutting over to App.getQueryParams in the selectors as it was difficult to disentangle from removing the configuration portion of the redux store.