Update all the things: frontend-base and frontend-build (#275)
* Update all packages.

* Unneeded packages removed.

In anticipation of frontend-base and frontend-build taking over.

* Moving card-validator into the src/payment/checkout directory

It’s only used there, so it makes sense for it to live with the code.

* Moving FormInput, FormSelect, and PageLoading into src/payment module.

… in appropriate places to their consumers.  The form input and select are in payment-form because that’s where they’re used.  PageLoading is up top.

* Moving utils and serviceUtils into payment module.

Renaming serviceUtils to handleRequestError since that’s the only thing in it.

* Using App.config for i18n configuration in tests.

Also doing the config in setupTest.js to remove it from most test files.

* Moving userAccount factory.

* Missed a deletion of configureI18n in prior commit.

* Using App.config instead of other methods of getting configuration.

There are a few I didn’t put in this commit because they’re on lines with other, unrelated changes.  Will merge them in next.

This commit also includes cutting over to App.getQueryParams in the selectors as it was difficult to disentangle from removing the configuration portion of the redux store.

* Cutting over to using App for config and apiClient

The whole app should now use App.config and App.apiClient.  Deleting environment.js in this commit.  This also ceases to call the “configureApiService” methods for most services except the top-level one, because it still has the interceptor.

* Using App.authenticatedUser and fetchUserAccount from frontend-base

Also cleaning up common seletors and actions.  common/selectors were moved into the order-details/data/selectors where they were still necessary, and the common/actions file is subsumed into frontend-base.

* Moving the response interceptor into its own file.

It will be attached to the apiClient in the top-level index.jsx file, rather than the payment service.

* Cleaning up PaymentPage test.

- The snapshot change is because of the updated version of font awesome from a prior commit.
- Changing its export from ConnectedPaymentPage to just PaymentPage.
- The “requirePaymentPageProps” thing in the PaymentPage.test.js file wasn’t actually doing anything.  At all.

* Creating a component called EcommerceRedirect

It’s used to catch 404s (in the MFE) and forward them on to the ecommerce service.  This will be used in our route configuration.

* APPLE_PAY_MERCHANT_IDENTIFIER is not a variable used in this MFE.

* Cleaning up styles and assets.

- Assets were not used by this MFE or are included in its dependencies.
- Normalizing the stylesheet and preparing it for using the new header and footer.
- Moving all the app-specific CSS into the payment/index.scss file - the top level one just glues everything together.

* Updating ADR to reflect some of the app configuration/apiClient changes.

* When using frontend-base, the redux store configuration files are no longer responsible for creating a history object.

* This is the culmination.  Refactors index.jsx to use frontend-base

Depends on previous commits being in place.  index.jsx simplifies massively and App.js (in this repo) is removed, along with the messages file which wasn’t being used.

- loadConfig is used to add additional environment variables.
- APP_AUTHENTICATED event is used to attach the interceptor.

* Use fedx-scripts and frontend-build config for jest configuration.

* Use frontend-build for linting.

And updating husky to its recommended configuration method.

* Let frontend-build manage babel configuration.

* Modernize travis config.

Also using fedx-scripts for i18n extract.

* Using frontend-build for dev server webpack config.

* Using frontend-build for dev-stage mode.

* Using frontend-build for prod build.

- Also adding “bugs” to package.json and updating the package-lock.
- The build.prod.js file has been renamed to create-apple-merchant-file and has been turned into a follow-up to the prod build, rather than a wrapper of it.

* Review feedback - cleaning up test code.

* refactor: moving speedcurve utils up to the top

Using them in index.jsx instead of the old “App” file since that’s the first moment we’re about to start painting.  Also updating references in OrderSummary and PaymentForm.

* fix: move speedcurve utils into payment so as not to break module boundaries

* Adding favicon.

* fix: usage of markPerformanceIfAble

* fix: send analytics events