fix: implement currency cookie detection as selector
We need to wait long enough for the app to know the name of the cookie via CURRENCY_COOKIE_NAME, which is attached to App.config during the startup sequence.  By reading the cookie during a selector instead of at load-time in the reducers file, we sufficiently delay that it picks up the real cookie value.

This means the currency reducer is no longer needed, as the selector reads it from the cookie itself.  That necessitates cleaning up a bunch of reducer/state related setup in our tests.