fix: receive 200s or 400s instead of 302s from ecommerce /checkout
We tried having Ecommerce send 302s to redirect the Payment MFE to the
receipt page. This didn't work because the library that we use for
requests on the frontend, Axios, doesn't appear to support not following
redirects.

This created a host of CORS and authentication issues due to logging
into Ecommerce from the Payment MFE domain.

We've changed ecommerce's /payment/stripe/checkout to send a HTTP 200 or
400 instead of a 302, along with the information needed for the
redirect. This commit instructs the frontend on what to do when
receiving these 200 or 400s.